pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quant_infra"
version = "0.1.0"
description = "量化因子研究核心库"
readme = "README.md"
requires-python = ">=3.12"
# 在这里列出你的项目必须安装的库
dependencies = [
    "pandas>=2.3.3",
    "numpy>=2.3.3",
    "joblib>=1.5.3",
    "matplotlib>=3.10.8",
    "duckdb>=1.5.0",
    "tqdm>=4.67.3",
    "tushare>=1.4.25",
    "scipy>=1.16.3",
    "python-dotenv>=1.1.1",
    "PyYAML>=6.0.2",
    "requests>=2.32.3",
]

[tool.setuptools.packages.find]
# 这告诉 setuptools 去 src 目录下寻找代码
where = ["src"]
